third_party/blink/web_tests/external/wpt/css/cssom-view/table-client-props-expected.txt

This is a testharness.js-based test.
[FAIL] Table with separated border
  assert_equals: Table with separated border clientLeft expected 0 but got 4
[FAIL] Table with collapsed border
  assert_equals: Table with collapsed border clientLeft expected 0 but got 4
Harness: the test ran to completion.
